backend/requirements.txt
fastapi
uvicorn[standard]
torch
numpy
transformers
Pillow
imagehash
pydantic
